visionOS: Enable multiview rendering
Enables multiview rendering on visionOS using layered rendering. For
this to work we also need to bump the version in our shader generator to
use MSL 2.1. On visionOS multiview rendering will be enabled by default,
the only way to disable it will be by setting the env variable
QT_QUICK3D_XR_DISABLE_MULTIVIEW to 1.

Pick-to: 6.8
Fixes: QTBUG-126294
Change-Id: I604595978c9555eebb13236ab6f4640949f60765
Reviewed-by: Andy Nichols <andy.nichols@qt.io>