XR: Make multiview rendering the default
Made multiview rendering the default for both OpenXR and visonOS when
supported. Devices that use OpenXR might not always support multiview,
in which case each view will be rendered separately as before. For
visionOS multiview support should always be available, but both
rendering paths are kept for now. This change also changes how multiview
rendering is disabled/enabled. Multiview is will no need to be
explicitly disable through an env variable and the ability to change it
through QML is disabled (Property should be made read-only in 6.9).

[ChangeLog][QtQuick3D.Xr] Multiview rendering is now enabled by default
on platforms that supports it.

Pick-to: 6.8
Change-Id: Ifaf6c04367ab61595223e1ed098a708f835e2283
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
Reviewed-by: Nicholas Bennett <nicholas.bennett@qt.io>
Reviewed-by: Elias Hautala <Elias.Hautala@qt.io>