XR: Make the property multiViewRenderingEnabled read-only
Toggling the multiview support isn't something we need to, and probably
shouldn't, support as it's not something that should be done outside
development and testing, which is why this is now controlled by setting
an env variable. The application will either do multiview rendering or
not for the lifetime of the application.

[ChangeLog][QtQuick3D.Xr] The multiViewRenderingEnabled property is now
read-only and toggeling multiview rendering is no longer possible at
run-time.

Change-Id: If99bc4c5d69da82557edee81d9261ff4b38ce90f
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
Reviewed-by: Elias Hautala <Elias.Hautala@qt.io>