CMake: Remove installation of bundled OpenXR headers
The 3rd party OpenXR headers are not used in any of the Qt OpenXR,
public headers, and no code actually links to OpenXRPrivate to try
and use them.
It was mostly a mistake that they were installed in the first place.

This fixes the issue on Linux where one of the OpenXR headers
'openxr_reflection_structs.h' is too big to be processed by syncqt
(specifically the std::regex_match implementation in GCC stdlib) and
causes a stack overflow crash.

Amends 7e6abbd1667c248adaccf2cf47afb4b712ec39d1

Pick-to: 6.8
Fixes: QTBUG-129807
Change-Id: I2610c1dacbfbe34801d099d186569d9ef8a4e040
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>