visionOS: Use new API to get the projection matrix
The recommended way to get the projection matrix is to use
cp_drawable_compute_projection() which was added for 2.0. This also
means we no longer need to calculate the projection manually or use the,
now deprecated, cp_view_get_tangents() function. This change also
changes the internal "eye" camera class so it's possible to set the
projection matrix directly, as we won't be setting individual components
anymore.

Task-number: QTBUG-129732
Pick-to: 6.8
Change-Id: I5eba8ab06031a29ed1547d4b37f67977b18bc513
Reviewed-by: Andy Nichols <andy.nichols@qt.io>