Ignore frameDelta when mouse rotating controllers
OrbitCameraController and WasdController are suffering from erratic
movement when rotating using a pointing device while the frame rate
changes drastically between frames.

The issue is that multiplying the mouse delta with the frameDelta will
make the movement bigger for slower frames and smaller for quicker
frames. This is now removed so the rotation is only dependent on the
delta of the pointing device.

To keep the movement speed as similar as possible to before, the delta
is multiplied with a constant factor so that it behaves identical as
before on a stable 60 fps system.

Change-Id: I5908b206727ca0453e8786112a8b4f2da3db8fb9
Reviewed-by: Andy Nichols <andy.nichols@qt.io>