volumeraycaster: Disable OpenMP on Android
The openmp library is found during compilation but not when running the
example on Android. Since it is an optional dependency that is just used
for faster conversion of the input data when available, it is disabled
for Android.

Fixes: QTBUG-130398
Change-Id: Ibcd4ba42e095d023a05a9cc7ecbaeab1573177c6
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
(cherry picked from commit cf838e2b0751cc7acd83c03110d15cec44c3a981)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>