Fix crash in the effect when used with singletons
This adds some extra checks to make sure QtQuick3D is set-up to
start rendering first. This check is already done in 6.x, so
this is just adding the same logic. The actual issue is however
not because the RCI is missing, but because the effect, or the
parent items, haven't been referenced by a window and therefore
a RCI cannot be retrieved. All the relevant code for this has
been completely rewritten for Qt 6, so this is a minimal,
relative low risk change for 5.15 only.

Fixes: QTBUG-125963
Change-Id: Ie180fac1e6122df46c63b0eef7170640c879adc3
Reviewed-by: Andy Nichols <andy.nichols@qt.io>