Fix broken shadow mapping with orthographic camera
The issue was that QSSGRenderCamera::calculateViewProjectionMatrix is
assuming the camera is a perspective camera. Instead of using this
method we use the original frustum points of the camera and move them
forward or backwards according to the split t values.

The same issue was present when calculating the radius of the frustum
when locked texels are active.

This change also makes the clipNear be the same as the camera's instead
of being hardcoded and clipFar not be limited.

This change also moves the pcfRadius expansion out of the frustum
calculation and does it on the shadowmap bounding box instead.

Fixes: QTBUG-132081
Pick-to: 6.8
Change-Id: Ic7cc725609946ab5adf7e23cf943e06c2a01b968
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
(cherry picked from commit e67c6bd6c86922bd833d9440a6c53d31380a5ed3)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>