Make module ready for source SBOM checking
This includes:
- turning VERIFY_SOURCE_SBOM ON
- adding rules to the licenseRule.json files
- correcting the licensing given via REUSE.toml files
- renaming license files not located in LICENSES folder.
Their name needs to be prefixed with `LICENSE.` to be ignored
by reuse and excluded from the source SBOM. The names are
updated in the corresponding qt_attribution.json

A lot of files are skipped during the license test,
but all are present in the source SBOM.
This is why corrections are needed before turning the
source SBOM check on.

[ChangeLog][Third-Party Code] Renamed certain license files outside of
LICENSES with `LICENSE.` prefix such that reuse will correctly ignore them.

Task-number: QTBUG-131434
Pick-to: 6.8
Change-Id: Ib0bad8e67e6d4f9a5d857376a11b4a2322940369
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
(cherry picked from commit 66b9c1e250e79fb0b517521a96b2927a3ccac95b)
Reviewed-by: Andy Nichols <andy.nichols@qt.io>