Overhaul material lights and shadows to be uniform driven
This change modifies how we generate materials with regard to lights and
shadows such that lighting and shadow information is distilled down into
a few shader key properties, instead of one per light/shadow combo. Now
all lighting and shadow code driven by uniform data so that it is not
necessary to switch shaders when the number of lights or shadows
changes.

The the case of lights, that means that instead of generating a bespoke
shader for each light, there is a for loop that iterates over the the
light uniform data.

In the case of shadows, all lights shadows are rendered into a single
shadow map atlas that can be used by sampling from a single texture. The
location of each lights shadow data can be looked up via uniform data to
know where and how to sample the shadow map atlas.

Change-Id: I1b76e6ce1d8334ba123596c551cd15d3b264fdea
Reviewed-by: Christian Strømme <christian.stromme@qt.io>