Simplify simple input example
Replace the ray pointer and the target with simple geometric shapes that
approximate the look. This is done to focus the documentation on how to
use the low-level input APIs in Qt Quick 3D.

The more advanced shader-based components will be put in a shared
folder by a later commit.

Change-Id: Idcabda9ddea79c0bfd6d83ad255e4d06fdf2d0d2
Reviewed-by: Elias Hautala <Elias.Hautala@qt.io>
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
(cherry picked from commit 80afa2c00e480756365e8bdeb2176dbeb02dddd2)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>