Use shadowmap index in texture name lookup
This only affects the name of the 'shadowMapTextureUniformName' variable
which is then used to lookup a cached binding for the shadowmap. Since
these are already unique when using the light index to get the texture
name this code works fine already, but we fix it to be consistent with
the texture name in the generated shader code.

Amends: aaaad943b30c68476f0dba352f38e2b1de56dd0b

Pick-to: 6.8 6.9
Change-Id: Iec85c72c997ea062a49f699501de1f7e17c4031c
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
Reviewed-by: Kristoffer Skau <kristoffer.skau@qt.io>