sampleNormalMap shader function should return a normalized value
It was assumed that since tanFrame was orthonormal that multiplying the
normalized value sampled would produce a unit-vector. This assumption
that tanFrame is orthonormal is not always true from a precision
standpoint, so it's safer to normalize after the multiplication.

This bug was having the side effect of causing qt_world_normal to be a
non-unit vector, which was problematic because much of our lighting code
assumes that the world normal value has already been normalized. This
should have only affected materials that used a normal map, as otherwise
the qt_world_normal is normalized before use.

The funcsampleNormalTexture.glsllib file has been cleaned up to remove
the versioning, as those were not being used and were intended to be
temporary.

Pick-to: 6.10 6.9 6.8
Change-Id: Icb15088abadbe36a23bdcac6946ab486c6bd46a6
Reviewed-by: Christian Strømme <christian.stromme@qt.io>