visionOS: Fix size update of the viewport
The internal viewport is normally updated when the size of the content
item is update, but since rendering is on a different thread on
visionOS, the update won't happen before the sync, which causes some
texture creations to fail as the target size from the viewport is not
valid (0 x 0). This was discovered while testing the extended scene
environment due to effects being used.

Pick-to: 6.8
Change-Id: I0c8f40c8cae291857c39c71c5f67e11dc392f11d
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 984b338448318bed138605cfdcb0442c0ffc599a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>