Make sure to generate tangents when only using a clearcoatNormalMap
To be able to sample any kind of normal map, we need to calculate the
tangent space, so for geometry without binormals and tangents, we need
to generate them. Previously we would use the uv coordinates from either
a bumpMap or normalMapImage to generate tangents, but in the situation
where you only have a clearcoatNormalMap, we just didn't generate
tangents at all, leading to incorrect normals being calculated when we
eventually sample that normal map. This code corrects this corner case.

Pick-to: 6.10 6.9 6.8
Change-Id: I4507abf0a3794d67e1db3a2c571ab4f9d5230106
Reviewed-by: Christian Strømme <christian.stromme@qt.io>