fix crash caused by static lambda closure capture
Using a closure in a static lambda is incorrect. This lambda does not
use the rhiCtx passed to the function. Instead, it captures the rhiCtx
that existed when the lambda was created and always uses it,
which may lead to unintended behavior and "use after free" errors.

Fixes: QTBUG-132042
Pick-to: 6.8
Change-Id: Ia868f009abbd75cf9de7b129662b0a2c2e37e5fe
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
(cherry picked from commit ef7b1d50535ba2f83608f5fbd2bcbe4dc181035e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>