Xr: Add some convenience functions for calling openxr functions
Not all features used are supported, or other wise accessible, which
means we need to check if the functions for each feature has indeed been
resolved. If a function cannot be resolved we will output a warning, but
that's not sufficient as we also need to check the function pointer before
calling it. Most place we already check this, but in the anchor manager we
don't and would therefore crash instead. With this fix we introduce some
simple helper functions to make this more convenient as it will include
the pointer check.

Pick-to: 6.9 6.8
Change-Id: Ifa59e82b345237823c50d6ca6c7d730499053089
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>