Make sure the built-in scene environment lives on the correct thread
This amends 361a1710fa89a37ee79f9a37398b321ba17ac1ef where we did not
take into consideration that the built-in environment could be created
on the render thread (environment() called during the sync).
This was caught by QObject thread checker, as we were setting the
parent object to be the scene root which lives on the GUI thread as it
should.

Pick-to: 6.9 6.8
Change-Id: I32395a31ec6b9901a600522cc57b8227dd75af1a
Reviewed-by: Andy Nichols <andy.nichols@qt.io>