XR examples: use Vulkan on Android
Android 15 deprecated OpenGL in favor of Vulkan. Vulkan was already the
recommended graphics API on Meta Quest. At the time Qt determines the
graphics API, we don't yet know whether the application is going be an
XR application. Given that it is too early to change the default for
all of Qt Quick, the best solution for now is to change our XR examples
to use Vulkan on Android.

Pick-to: 6.9
Change-Id: Ie493879911969b7bcbcef48371008bfae0725b9c
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>