Assimp: baseColorFactor is linear and needs to be sRGB
glTF2 models that used baseColorFactor could end up with the wrong color
when imported because we expect all "color" type values in Qt Quick to
be sRGB colors, but in the case of baseColorFactor this value is already
a linear color value, so it got unnecessarily converted. So for this to
work correctly, we need to make sure to convert this value from linear
sRGB on import.

[ChangeLog][Quick3d][Important Behavior Changes] glTF2 Models that use
baseColorFactor now work according to spec, but will be differnt than
before.

Change-Id: Id85fa6b63b8087adfad2a66d13d551d2458e23c5
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>