visionOS: Fix teardown when running in the simulator
Three cases were identified while testing the baseline test. The fist
cases happens when the application is aborted before the scene, the
graphics resources, and the internal window has been created, causing the tear-down to soft assert.
The second issue happens due to the simulator running in single threaded
mode and we attempted to shut down the render thread, which isn't created
or used in the simulator. The third and last case is the AR session
state messaging coming in on a different thread after we've already shut
down the session, this is fixed by changing the message queue before
stopping the session.

Pick-to: 6.9 6.8
Change-Id: I7b61bb50dd191b22f65c15281d222cd2a81ecde0
Reviewed-by: Andy Nichols <andy.nichols@qt.io>