visionOS: emit the frameReady signal when running in the simulator
We relay on the frameReady signal for our baseline testing to grab the
content of the screen. Unlike OpenXR we'll just enable this for when we
are built for running the simulator, as that's when we're actually going
to use it for now.

Change-Id: I93128060a5de658894caf2208c704628cba5c5b7
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 6c93c91a2edcc5213a91c088b5d2a6075b67e893)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>