XR: Small improvement to the scenegrabber
Move the elapsed timer to the outer scope to avoid that we hold onto a
reference to it in lambdas that can outlive the current scope.
Also added the logging category to the developer option as a convenience
when diagnosing issues.

Change-Id: Ib2bac02aeb702a7b7756502f65ce1c65c42b0ea8
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit a8ef96c116a678d8e3c13b745cd8ae2716cb5c78)
Reviewed-by: Nicholas Bennett <nicholas.bennett@qt.io>