visionOS: Ensure that the layer renderer isn't released underneath us
This fixes instabilities seen when an application is suspended/exited
due to the layer renderer being released prematurely. Since the layer
renderer is our primary access to the state of the rendering we need to
keep it alive at least long enough for us to check the state.

Pick-to: 6.8
Change-Id: I6413bdedc63b38a140aa1a713341297e61ab7c66
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 931784fabd44ea555d7c5481b82a665da563a471)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>