Add QQuick3DTextureProviderExtension API
The QQuick3DTextureProviderExtension is a specialization of the
QQuick3DRenderExtension class that is used to create a custom texture
provider extension that can both provide additional metadata about what
type of texture will be provided and also automatically register the
extension with the QtQuick3D scene graph. This means it is unecessary to
manually add the extension to the list of extensions to be used with a
View3D, and just using the textureProvider property of the Texture
component is enough to trigger the extension code to be run when
necessary.

The proceduraltexture example has been expanded to include a
demonstration of creating a texture with a GPU pass.

Change-Id: Ic366ffb625a4fb0f5540edba5b9c0bc19123e870
Reviewed-by: Christian Strømme <christian.stromme@qt.io>