Lightmapping: Structure change
This change contains a refactor to make requesting and the
initialization of baking more coherent. It lays the foundation for the
patches that comes after.

All request handling is now only done in
QQuick3DSceneRenderer::synchronize, instead of the env/cmd variant being
handled in-line in the layer at the beginning of the render loop.
QSSGRenderer::rhiPrepare is way too late to do any kind of setup, so
unifying them in the synchronization phase is important for consistency.

QQuick3DSceneRenderer then passes on some data to the layer, which holds
an instance of a new class called QSSGLightmapBaker. This is an
indirection to isolate the lightmapper a bit more, letting lightmap
baking scale better and make testing and debugging easier.

Change-Id: I0691964bc753b8227f5c3147a9b27699448f6405
Reviewed-by: Andy Nichols <andy.nichols@qt.io>