Expose delegateModelAccess from QQuick3DRepeater
It mirrors the same property from any internal delegate model. If the
internal model is not a delegate model, the value is "Qt5ReadWrite".

[ChangeLog][Quick3D] Repeater3D now has a new property
delegateModelAccess. Setting it to DelegateModel.ReadWrite allows you to
write values into the model via required properties just as you could
with context properties.

Task-number: QTBUG-132420
Change-Id: Ie49a08f7f87cf2d8ee5b28d2860bf4785e4cce81
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>