QQuick3DRepeater: Fix model creation
In setDelegate(), we want to create the model if it doesn't exist, not
if it's not ours. Otherwise we might overwrite a perfectly fine model
set by user code when creating a delegate.

Also warn when we cannot retain an explicitly set delegate because an
incompatible model was passed.

Adapted from commit 50c616c5605a5a223d058d69336dc82d48040e98 in
qtdeclarative.

Pick-to: 6.9 6.8
Change-Id: Ic7af11155f955b0ee5037a304f496c7aa9aea822
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>