Lightmapping: add custom lightmap file format
The EXR file format is a bit slow on multipart files due to the fact
that it has to decode all headers to find the header you are looking
for. Therefore we add a custom file format designed to be fast and
easy for storing baked lightmaps.

The file format is basically just a sorted list of key/value pairs where
the keys have a fixed size with offsets and sizes into the file. This
makes it possible to lookup keys in log(N) time and you don't need to
parse the whole file or all the keys when looking for a key.

There are helper functions for storing f32 and u32 images, meshes and
metadata.

The format is storing data little-endian but will convert to big-endian
automatically.

Change-Id: Ia230b98b38b685a44022df66a00b42ca9f7ca39b
Reviewed-by: Kristoffer Skau <kristoffer.skau@qt.io>