Lightmapping: Add NLM GPU denoising support
This change adds support for using Non-local Means (NLM) GPU-based
denoising. The denoising can be called either through running the
application with --denoise-lightmaps or by clicking the Denoise
lightmaps button in the DebugView.

The way lightmap baking is done is changed a bit. The direct and
indirect lightmap is separated into two lightmaps and are merged at the
end of the process. Only the indirect lightmap is denoised. The direct
lightmap is instead supersampled, gaussian blurred and downscaled to
reduce jaggies. Since the direct light is already deterministic,
denoising it will only add blurriness and remove features.

The aggresiveness of the denoising is controlled with the
Lightmapper::denoiseSigma property, going from 0..100.

Since the unwrapped uv picture does not guarantee neighbouring pixels
are actually neighbouring in the mesh and taking these into account
would lead to artifacts, the denoising algorithm uses a mask so that
only pixels that are part of the same chart will be considered
when denoisinig.

[ChangeLog] Add support for denoising of baked lightmaps. Add
Lightmapper::denoiseSigma property. Add 'Denoise lightmap' button to
DebugView.

Change-Id: Ie009c20c4a71c25c9b92eaa9125d15afb1ed0b92
Reviewed-by: Kristoffer Skau <kristoffer.skau@qt.io>