Don't insert null bone texture for each model context
This was clearly a mistake introduced in 4d066e99bdc43ff96513343f5f335a3676277c8d,
as the original code updated the pointer on the model
context regardless, and without cost. Since we now
store the texture in a map there's absolutely no good
reason why we should insert a null texture for all
model contexts.

Pick-to: 6.10 6.9 6.8
Change-Id: I7321218161e0037a24adbad6487819f54e59765d
Reviewed-by: Jonas Karlsson <jonas.karlsson@qt.io>