Fix issue with the scene root node being removed twice
The scene root node is created by a front-end object and manually added
to the layer, that means the scene manager will release the back-end
object as normal, but since we added it to the layer manually it will
never be clear out of the layers child list. This hasn't been an issue
before now, as both were released at the same time. With the
introduction of the "global" root node, all nodes under a sub-tree
(layer) will hold an indirect reference to the "global" root node
through the layer, so we delay destruction of the layer until those
nodes are released. Due to this delayed destruction the layer could
still be marked dirty, which eventually would cause it to mark it child
nodes dirty as well.

Added some extra checks and removed the automatic removal of the layer
from the root node, as that should be an explicit call.

Fixes: QTBUG-137479
Change-Id: I28e5ebd72dd572ebeec0db5997d88731bdbd6b94
Reviewed-by: Sami Varanka <sami.varanka@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 95ae6b65f8c0d05b71395007b438bfca8d19b057)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>