Fix getModelBounds() for lightmapped models
QSSGBufferManager::getModelBounds() now checks if the model has a
lightmap and returns the bounds of the baked mesh in that case.

Since loading a lightmapped mesh requires the path to the lightmaps file
to be set, and since the render layer might not exist when
getModelBounds() is called, the lightmapper options are now stored in
the scene renderer instead of the render layer.

Pick-to: 6.10
Change-Id: I85c8998095994ba416b1fd7c7b77212c9dafb4e1
Reviewed-by: Christian Strømme <christian.stromme@qt.io>