Fix lightmap file warnings
While the previous patch suppressed the initial warnings as expected, it introduced a bunch more. This is because the code in question runs every frame, which was missed. Also the fact that lightmaps are implicitly enabled at all times even if a Lightmapper is not explicitly created by a user. So the check needs to be handled in the places where we end up querying lightmap data, which is the only way for us to really know if the user truly cares about the lightmap file being present or not.

Amends: f7ed5317806b59e50369e63461331cec02670e70

Pick-to: 6.10
Change-Id: Id31183036f282d32905ba4ddd0ed7baa9439b61f
Reviewed-by: Jonas Karlsson <jonas.karlsson@qt.io>