Lightmapper: Render direct light as tiles
Previously, the direct light was rendered to a x3 sized texture that got
blurred and downscaled to get smooth looking edges. The drawback with
that method is that it needs 9x the amount of RAM and VRAM. Instead, we
now render this upscaled texture in smaller tiles.

Four floats for defining the UV region is added to the lightmap
rasterizer shader. These floats are then setup for the tiles so that
they are rendered into a subregion but with higher resolution. These
tiles are then blurred and downscaled and put into the direct light
texture map.

Pick-to: 6.10
Change-Id: I63484dc0856f8e4ee7a9fc43dd1b9b7a2d1b3506
Reviewed-by: Kristoffer Skau <kristoffer.skau@qt.io>