Lightmapper: Fix when we set models to bake
Currently we set the baked lighting models on the lightmapper prior to
triggering the extra frame that's necessary. This is not safe. Instead
query the lighting models when we are on the baking frame, right before
starting. Added a new callback to call the necessary function on the
layer.

Amends 780ab6e03190710dd921425b37afae5a12a36b63

Pick-to: 6.10
Change-Id: I71475d3d55b2f1b49557c4bf2df67097b3928e6b
Reviewed-by: Jonas Karlsson <jonas.karlsson@qt.io>