Add lightmapviewer tool
This is a debugging tool for viewing Qt's baked lightmap files. While
users can use this tool, it is intended for development only.

The tool is both a GUI tool for viewing lightmaps and a command line
application that accepts flags for printing info and extracting the
content of the lightmap file.

Pick-to: 6.10
Change-Id: I95279e248ea82e0e1384bbf880dcecc7af778dfe
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>