Expose QQuick3dRenderStats to QML
The type is accessible through the `renderStats` property of the exposed
View3D type but is not exposed itself.
qmllint produces a warning about this as it cannot resolve the type that
is being accessed.

Thus expose the type to the QML type system through an appropriate
registration macro, in particularly naming the resulting type
`RenderStats` to align it with the pre-existing documentation for it.
Additionally marking it as QML_UNCREATABLE due to the already existing
documentation for the type stating that the type cannot be created
directly but only retrieved from a View3D.

Additionally register the version of introduction for documentation
purposes.

Pick-to: 6.10
Task-number: QTBUG-138170
Change-Id: Ifba348924815cd4457c653dae1f8dc6ebf188756
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>