Doc: Fix incorrect usage of \generatelist
The \generatelist command takes an optional argument
that affect how the members are listed, for example,
`classesbymodule` which creates an annotated list of
the classes in a named C++ module.

`groupsbymodule` is not recognized or documented. QDoc
falls back to generating a list of all members of a
group in this case. While the end result is correct,
this relies on an undocumented feature that is likely
to change and break.

Replace these instances with an explicit \annotatedlist
command that produces the same output.

Task-number: QTBUG-138901
Change-Id: Idc09856c99cc49a5909cb86684e7e18a4fe84190
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 18760dc6a4fb97202784c3d5e1e453873ef833bf)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>