Add RenderOutputProvider helper
This component enables backend texture to be used on the frontend via
the textureProvider API of the Texture item.

The API is for now limited to a minimal set of the render results
from some of our internal passes. Somewhere down the road the
intention is that a possible user pass type will be able to expose
its result(s) through this type as well.

Change-Id: Ib62f7dd34fb68cf7d6f6c261e1106582ed641ea8
Reviewed-by: Andy Nichols <andy.nichols@qt.io>