Remove unusable enums from QSSGFrameData
These were technically added in 6.9, but they were not usable in
practice and it doesn't make a whole lot of sense exposing them as is.
Since these cannot possible have been used outside QtQuick3D we should
remove them until we know how we want to expose them. The enums values
are still available internally and can still be reached from the
outside, on the off chance there are usages in the wild.

[ChangeLog][Quick3d][Important Behavior Changes] Removed RenderResult
enums in QFrameData that was only meant for internal usage.

Change-Id: I40ecf4031bde5f30475935c47b9de07280cee235
Reviewed-by: Andy Nichols <andy.nichols@qt.io>