Don't automagically release registered texture providers
Just do what the documentation says, register the texture provider and
keep it until the caller unregisters it. This change also removes the
memory stat for extension textures as we'll need a slight different
approach then what we do for the other resources - We were increasing
based on usage, which isn't correct.

Pick-to: 6.10 6.9 6.8
Change-Id: I7dd094bfddb3fc4cb02ca07c011c131e8dc8ba0c
Reviewed-by: Andy Nichols <andy.nichols@qt.io>