QQuick3DXrAnchorManager: make QQuaternion construction constexpr
The QQuaternion::fromEulerAngles() function is out-of-line, so cannot
be constexpr. But we don't need to call it to construct a static const
object of a trivial rotation; we can just create the quaternion
directly.

This turns a mutex-protected run-time ("magic statics") initialization
into a compile-time constant.

It also fixes a compile-error that the introduction of the new
fromEulerAngles(EulerAngles) caused, due to the call here using
untyped braced initialization (even though just dropping the braces
would have been shorter).

Amends 41aa8aa2d47ab710dd48b536e58bd6080feafe1e.

Pick-to: 6.10 6.9 6.8
Change-Id: I62e98062aaa2d596342005d337328d3b790d600d
Reviewed-by: Christian Strømme <christian.stromme@qt.io>