Fix quick3d frame view in qtcreator qml profiler
The quick3d frame view is empty, because it's missing required
information in the profiling events.

Set the missing profiling id for the scene renderer so that it can
forward it to the profiling events.

Also adds autotests for the profiling.

After adding autotest a crash was observed in QThreadStorage. After
investigation is seems that the QThreadStorage needs to be a static
variable so changed that.

Fixes: QTBUG-139324
Pick-to: 6.10 6.9 6.8 6.5
Change-Id: Id3d0401a794792e5c03290afa64142be1376ca96
Reviewed-by: Andy Nichols <andy.nichols@qt.io>