Correct scene normals for instanced picking
QQuick3DPickResult::sceneNormal() was calculated from the local normal
and the scene transform. With instanced rendering, this needs to be
calculated based on information that is not available on the front end.
Therefore, move the calculation of the scene normal to the picking
implementation, and remove the trick where m_normal could either be in
local or scene space.

Change-Id: I71a8bcf9ae22960c2136e30e59682fcaeea3db57
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 2b827a27720591d8786dd8ee0a27eabd1dd7e678)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>