Fix Item2D data storage
Store data per-layer for Item2Ds. Since an Item2D can be shared between
scenes we got into the situation that the data got corrupted by one of
the scenes update (which ever update the item data last). With this
change we now store the data for the Item2Ds in a separate storage,
which is per layer. This causes some complication when items are
destroyed, so for the time being the Item2D is maintained by QtQuick3Ds
render class. This fix is also meant to be compatible with versions
before 6.10, as a follow-up change will be needed to unify the storage
approach similarly to how it's done for models and node data.

Fixes: QTBUG-136911
Fixes: QTBUG-126098
Pick-to: 6.10 6.9 6.8
Change-Id: I35da0d0fc1a35832d78c4f6c772d95037edc81a2
Reviewed-by: Inho Lee <inho.lee@qt.io>